Cleanup text watcher logic.
Adds specific text watchers designed to look for real
chagnes to the to/cc/bcc fields and not just when a chip has been
matched with a contact in the user's database.
Also fixes a bug in comparing selected accounts to compare the
account names (duh!) instead of the name to an account object.
moves visibility of quoted text call to the quoted text view as
the calling app shouldn't have to know about its visibility to
add text. Adding quoted text should be enough to make the view
visible.

Change-Id: I7fbddcd1cb69672e7a07229bf2d60bde36512220