Make sure we return the correct view for getView.
getView should only provide the view we wish to show at
the top of the screen. Since we are tracking this, just
always return the same view.
It would also appear there are bugs in getItemViewType,
so I am pulling this and putting a note in the UI spreadsheet
to make this more effecient later.
This basically disables the recycler for now.
Change-Id: I878c4cc99a6dd6468b2c83ad0c9e8bb23d7ddc03