Add bits of Attachment info to message cursor results
Ultimately, message view needs more information about how many
attachments a message has and what sort of files they are to
render UI upfront and allocate space in the HTML for them.

I'm adding a new field in Message cursors to contain this basic
attachment info as a JSON array. Soon we can expand the format
to be useful to compose, and can deprecate
'joinedAttachmentInfos'.

Change-Id: I9c264a6f09b376f5ee9b9f56e11fa80dc5faec56