Ignore first time non-user navigation event.
Going from widget to conversation view doesn't give the list
navigation spinner a chance to come up. We have a current account
(mAccount != null), since we are viewing a conversation. When we
transition back to list mode, the spinner is brought up for the first
time. Since the spinner doesn't take a starting offset, it chooses an
offset of 0. This fires an onNavigationItemSelected(position = 0),
which resets mAccount and the account is changed everywhere.

When the list navigation spinner is brought up for the first time and
we have a valid account, it isn't a user event. Ignore it.

Fix b/6379319

Change-Id: I60ac51b096f1924018f21be81d9a26f7b6861dfd