Clear the ime suggestions/ underline for suggestion when saving.
Otherwise, we save with an underline if the user hasn't committed
an autocomplete
Executive decision: we use to only auto clear it if the user tapped
back and we-auto saved, but if the user taps "save", they probably don't
want the underline either...so I'm clearing it in response to save draft
as well.
I also remove the underline from text in the subject, as that would
be equally annoying to save.
Change-Id: If5e2a09bf0028e54cc7763463dcd5dcf75369361