Attachments now show in a grid-like view.
Currently, two per row are displayed in portrait
and three per row in landscape. Need UX decision
on how many to show for larger screen sizes.

Much of this code is taken from Contacts which
used a similar layout for favorite contacts.
That layout was taken and tweaked to suit the
attachment handling purposes.

Still need to add the appropriate interactions
for the various attachment types per UX's guidelines.
Will add that in the next CL.

Change-Id: Iad6ea716a680b1fd48f16d34de2a5bc2e87b7db5