Reworked attachment grid to use a dp-based layout.
Instead of having a resource selection for the
number of columns (which requires
a different value for portrait and landscape as well
as every screen size), utilize a minimum dip tile size
that figures out, while measuring, the appropriate number
of columns to use and expands the tile size to fit the
entire width with that number of columns.
As a result we have only two different values
for the minimum tile size. 120dip is the default and
180dip for sw600dp. Verified that the appropriate
number of columns occurs on all Google experience
devices.

Additionally, added a header and bar for the attachments
so that the user can clearly see where attachments
begin.

Change-Id: I48b1645e1004596747c1585e44a387b1231bd661