Non-image attachments now display as a list.
There are now two separate attachment areas in
the MessageFooterView. The AttachmentTileGrid
contains all images in the gridded tile look.
Additionally, there is a LinearLayout that contains
0 or more MessageAttachmentBars. These bars contain
a small icon, info about each attachment, and
an overflow button that contains additional options
other than the default for that attachment type.

In addition, there is now AttachmentActionHandler
that performs a lot of the interaction that both
attachment views share.

Change-Id: I60d297b7514e6ee35d23a9ec250833e886328d4d