UI improvements to photo viewer.
Photo viewer now shows the display name of the
image. Additionally, it now has a subtitle that
contains the current number of that attachment.

Additionally, the implementation in Unified utilizes
a derived class of PhotoViewActivity that tweaks
the subtitle to display image size and adds action
items. Currently the items are visible but do not do
anything. Next commit will wire these up.

As part of refactoring to make the derived classing work
properly, a bunch of legacy code was deleted including
all of BaseFragmentActivity as its sole reason remaining
was for spoofing the ActionBar.

Finally, a small change to the cancel attachment button
was made so that it was appropriately selectable and
48dip by 48dip.

Change-Id: I5a50bd420fa23a3e7d8734b5a46a6f28198ccef4