Show wide conversation list item on wide tablet
Show the wide conversation list item in conversation list mode when running
on a device where the current width is >=1000dp.

Since on a 10" tablet we are supposed to show different conversation list
item layouts depending on the current width, this resource specified is
using the wXXXdp qualifier.  If we were to use the sw720dp-port qualifier,
a tablet that is 720x720dp, would show the wrong layout in one of the
orientations

Bug: 6600648
Change-Id: I0388bce4c0cf8bac5a6a5bb9cbf126aab113ef69