Compose now shows images as tiled attachments.
It's still a bit buggy (notably, once the previews
do not show up after the initial attaching). Those bugs
will be fixed in the next commit but this was a good stopping point.

In order to reuse as much code as possible, the AttachmentTileGrid
was moved into com.android.mail.ui. Additionally, MessageAttachmentTile
was split into AttachmentTile and MessageAttachmentTile. AttachmentTile
is a base class that handles all of the tile interaction that is common
between the browse and compose versions. At present, this is mostly
the fetching and displaying of the preview image.

As a result, MessageAttachmentTile now derives from AttachmentTile.
MessageAttachmentTile handles the interaction specific to browsing
messages (such as opening the photo viewer, etc).
There is also new class called ComposeAttachmentTile
which derives from AttachmentTile and handles the interaction
specific to composition of messages.

The AttachmentsView in composition now contains two children in its
view hierarchy, an AttachmentTileGrid and an inner LinearLayout
that now stores the non-tiled attachments.

Change-Id: Iad85529a7b6d573060645499271860cf07877754