Collapsing attachments in compose now works.
When the draft has attachments, if the focus
on any of the text fields changes, the attachment
area collapses into a smaller area. Tapping that area
causes the attachments to expand into the full view
with the previews. Some refactoring was necessary to
enable reuse of the thumbnail loading for the small
preview icon.

Change-Id: Iaefb71736a2359b9090b2fe736ad2f33a9838ed7