Fixed bug where photo jumped at end of scale.
At the end of the scale gesture, if one finger was
removed before the other, an onScroll gesture would get
called, causing the photo to jump to where the remaining
finger was pressed. Added a debounce flag so that
the first onScroll event would be ignored. Seems to
fix the issue.

Also, deleted some unnecessary code from PhotoView.

Change-Id: I01dc069a9dbfd6c5eea13e373bbfb0fea492c857