Undo Bar refactoring and layout fixes.
Previously, the undo bar was declared in xml
and included a layout called undo_bar. undo_bar
contained merely a LinearLayout and an include
for undo_row. Since the undo bar was just a FrameLayout,
turning it into a LinearLayout enabled the elimination of
undo_bar.xml.

In testing this change, noticed several bugs in the layout
of the undo bar. It was not properly accounting for left
and right margins in two pane mode, was obscuring the
pager strip in the conversation view, and was not showing
up in the left column when performing a batch operation
in two pane mode. All of these have been fixed.

Change-Id: I0c6a2656c5a86d38aab6fef074b16f1b61f20a1b