Improved performance when scrolling conversations.
Previously, every time the MessageHeaderItem was bound
to the MessageFooterView, we removed all of the internal
attachment views and set all views' visibility to GONE.

With this change, we check to see if the attachmentsListUri
has changed. If it has not, we skip the reset step described
above. That way, the only time we remove all of the views is
when the attachmentsListUri changes.

Change-Id: I93d9008f95b28a580b4e628361b0030b81e286a9