Added the via domain to the UI.
When the via domain exists and the message is
expanded, the message subtitle shows "via blah.com"
instead of the normal address. Additionally, if
the via domain exists and the message details is
expanded, a from row appears that shows the full
name, address, and via domain. This row only appears
if there is a via domain.

Change-Id: Ic39d7d90d72592e300fc6d6144e62e4b08db3b91