Fix widget updating
Changes that happened by the user, from the UI, were not causing
the widget to get updated.  This was because the UI provider contract
is to suppress notifications that where performed by the user
from within the app.

Now a client of the UI provider can specify that they want all
notifications

Bug: 6870004

Change-Id: Iae0e27fd8b4dca7cd690f3b898d3861b38d92f95