We weren't using the cached header models correctly.
This should speed up scrolling since we allocate fewer objects.
Also, fixes the issue where we were not properly creating a new layout
when the read state of a message changed in the case where a user opened
an unread message and then went back. Now, all the senders are properly
bolded/ unbolded.

Removed the deprecated conversationInfo.sendersInfo, which was only being used
by gmail, and had been replaced by using the proper messageinfos
Also, performed some surgery on the widgetviews so that it acutally
uses the messageInfo from conversation info

Change-Id: I43fde9cd23ebc12f2de68e0d18b941304b4a2be7