Fix viewing search results
When a setting is instantiated, sometimes a missing signature
is causes the signature memeber variable to be null, and it appears
that some times it is an empty string.  TextUtils.equals() returns false
if only one is null.

Bug 6894588 has been written to track how to represent unset settings
values

Change-Id: I6a756da3ebfdc1e413aa4aac6ce91055380022b1