When we update the folder the user is viewing, clear the items cache.
When we swap folders, a lot of information about an item is going to
change AND we wont be seeing a lot of the same items.
It makes sense to just blow away the cache and redo.
fixes b/6987220 User defined label added to a message is missing in Gmail Inbox after reading the message
Change-Id: I1167af604bab4045946c89bcacf3f87bd2077919