Don't show any folder colors that match the default background color.
This is a signal that it should just blend with the background.
Fixes b/6819641 White artifacts on the label list
I added a bug for a better / clearer fix at some point in the future
When we redo the gmail provider, we can clean up how we propagate
that a label doesn't have a background color.
Change-Id: I8812da1ab5799791b453e89db4262c12e0228fd7