Fix resetting ConversationCursor
We need to reset the state of the Activity controller when the
ConversationCursor is reset.

It isn't necessary for ConversationCursorLoader to call the
callback's onReset, as this will happen from the loader manager
Also, the loader manager only calls onLoaderReset() if the caller has
valid data from a new instance of the loader, but onReset of the loader
is still called.  We don't want to call onLoaderReset() for every call
to onReset()

Bug: 7009556
Change-Id: If943409028aa8f5e465a5dba4bac1307a4d4e0fe