Turn on Andy's most excellent copious logging of layouts
When the 'MailBlankFragment' logging is enabled, this will disable all
hacky workarounds and log view state. When a blank view happens, look
for logging with this tag to see where the view state might have
messed up.

With the 'MailBlankFragment' tag disabled, this removes all logging
and enables the hacky workaround to patch-up view layout requested
state.

Bug: 6946182
Change-Id: Ia4c2515f969be29678dc0a9bd1c11d487e3d1520