Serialize the account just like we do for ALL THE OTHER CASES
we always put account in extras in a serialized format elsewhere

Do it here, too
Otherwise, it comes in serialized (json) but we look for it as a parcelable
but the upgrade serializes it...too confusing! Just use the same method everywhere.

Fixes b/7013037 Compose new email shortcut on Gmail widget incorrectly populates from and to

Change-Id: I3bb88f542fe802cdfcbbe3da1eb8207602110d44