fix scrap view reference loss that caused native crash
When re-rendering an existing conversation view, the part of the
measure process that uses scrap views inadvertently removed them
from the scrap heap without first calling removeDetachedView().
This made the views eligible for GC, and if a GC occurred before
the next draw, ultimately led to a native crash.

Fix this by NOT removing scrap views during the measure phase.

Bug: 6905156
Change-Id: I7ea4b2ec48ce683cbbf0d40247ae5040c777854b