Fix Hiroshi bug and allow back from folders other than inbox.
If mInbox is null, we have never created an Inbox before, and in this
case it is indeed correct to start a loader for the default account
inbox. When AAC.onLoadFinished() is called, it will return a cursor
for the conversations in the default inbox.

This account will be triggered if you directly enter a non-inbox
folder directly from a notification/widget without entering the inbox.

I was wrong in thinking that the loadAccountInbox() codepath was
incorrect.

Change-Id: Ie1a776b1e524432bf9f6700ef5291c0dc0ff0917