avoid doing fragment work during pager swipe
Pager swipe was triggering immediate fragment transactions via
mark-read, so just don't call onConversationSeen so early.

All that really needs to happen immediately when swiping is:
a) subject-in-actionbar should be updated, and b) conversation
list highlight should change. Do both of those now in
AAC.setCurrentConversation, and do only that.

Bug: 6976876
Change-Id: I1f679627b36d25faabc27eb7dc9dac191aab1d7b