Conv. list CHOICE_SINGLE, highlights conversations from widget launch
The problem lay in the complex timing dependency between obtaining a
ConversationListFragment (not available immediately after a Fragment
Transaction), setting conversation.position (the tracker was only
calculating the position on requests to next/previous) and updating
the CLF on cursor change (the call to AAC.getCLF failed on the first
conversation cursor load).

The solution is for the CLF to rely on mConversationListObservable to
obtain cursor changes, for the position tracker to calculate the
position when initialized with a conversation, and for the CLF to set
the list's selected position when the cursor is updated (if it hasn't
been set already).

Also, the CLF uses selection to highlight, and CHOICE_MODE_MULTIPLE is
the incorrect option. Using CHOICE_MODE_SINGLE allows us to avoid
calling clearChoices() in CLF.setSelected(int position).

Partial fix for:
b/7145111 item not highlighted in inbox list when open message from notification

Change-Id: I955d83b44b471ca9d4424e779f854bb664036833