Don't show two archive icons
Since both R.id.archive and R.id.change_folder correspond to the same
action, they have the same drawable: drawable/archive.  Only one of
these is ever selected, based on whether we are in the inbox or in a
user label. The selection is done in onPrepareOptionsMenu, and one of
them is disabled. However, in the time between when the menu is
loaded, and when the actions are selected, the user can see two
archive buttons briefly. The solution is to remove one of them,
doesn't matter which one.

This fixes the issue of showing two archive buttons, but still has a
problem: When visiting a non-Inbox label for the first time, an
archive button is shown in the ActionBar and then hidden.

Fix b/7161358 Duplicate "Archive" icon displayed temporarily when an
    email is loading in the Gmail Android app

Change-Id: Ibb9e0291ae8ca73063c4de80c37aa5aa894f54fa