Explicit Intent to view Inbox
The Compose Activity was creating an intent with a folder object
populated with uninitialized values.  Using this object is very
unsafe. Since the entire goal was to view the default inbox, that is
easily done by passing an intent without a folder. AAC.handleIntent()
shows the default folder when only an account is specified in the
intent.

Since using new Folder() is very unsafe, the method is made private,
and a static method with a suitably dangerous name was created to
ensure that callers are aware of the peril involved.

Fixes Bug: 7139995 Empty conversation list

Change-Id: Ic3e254630afd59a59caba10f2249e3e451005d63