Not in conversation view -> unset current conversation
When the current conversation goes out of visibility, we should unset
the mCurrentConversation because many objects get the conversation
from AAC.getCurrentConversation().

I tested this with both one pane and two pane, with search and
otherwise. But I wouldn't be surprised if this breaks something. If it
does, revert!

Fixes Bug: 7205219 Message list is scrolled down by itself until it
           reaches last read message

Change-Id: I263c07e8063e7094a1defca76dde788b52df71ff