Remove the current folder when dragging to a new folder.
Drag/drop should remove the currently viewed folder as well as
add the target folder.
Add a type for "all_mail" so that we don't remove the all mail folder, since
that is a special provider type.
Fixes b/7216571 JB MR1 tablet: drag n drop does not remove the label where the msg was dragged out of

Change-Id: I80061f9ee4f58c9ff24b290662f46724c89e61f2