Second fix for attachment reliablity
If an APK had been pre-cached, we would fail to launch the package
installer.  This is because the package installer's intent filter
only handles file:// uris

Now if the UI has specified that the destination is external, it will
make sure that the attachment has been saved on the external storage
before sending the view intent.

Bug: 7133835

Change-Id: I408dc70f1d12a65e15535280484b59cf1a6e810a