cherrypick: Populate the minimum fields for a folder.
The others are a waste for the folder list fragment.
Make it possible to create a full folder object from the
cursor associated with the folderlistfragment when necessary.
Part of b/7301434 (*) gmail extremely janky on manta

makes a significant dent in jank: the operation took
between 100-400ms (depending on label list size)
now its < 25ms
Change-Id: I5dbd5b34a4708d882b7ea2bfd6e7675b1350d6ff

Conflicts:

	src/com/android/mail/ui/FolderListFragment.java