When updating a message body, don't collapse quoted text
When a reply is sent and synced, the quoted text portion
transitions from an expanded snippet to a collapsed one,
which shrinks the conversation height and causes the scroll
position to jump.

It's not enough to just avoid processing reply bodies, because
a reply to a wide message must also be normalized if the quoted
text portion is visible.

So when replacing a message body, switch the initial state of
quoted text blocks from collapsed to expanded.

Bug: 7246054
Change-Id: I5375beeb4209779e7cb7feaa570590a9cbc520da