Defer folderlistfragment updates while we are animating.
Part of b/7301434 gmail extremely janky on manta

this will have no effect on phones as we never show
folderlistfragment at the same time as convlistfragment,
so we can never be animating!
this will also never trigger unless the user causes an explicit
delete of an item, since only that triggers animations.

Change-Id: I27cf0542897e364212ea687e2eef9a424a2090a5