Don't reuse the leavebehinditem.
This item should be recreated if its of the wrong type rather than
getting nulled out and reused. It would appear that the recycler
was returning views that were still in use.

Need to determine root cause but this certainly masks the issue.

Fixes b/7305963 Undo archiving is very slow in Gmail

Change-Id: I4481a27bbc02dd9419d21a77268a8c08d0059b3c