Enabling recent folders? Request them too.
On phone, entering conversation mode disables recent folders. When we
re-enter conversation list mode, we enable the recent folders, but
never update the recent folder list. As a result, there is a window of
critical time between:

enableRecentFolders()
... and ...
mSpinnerRecentFolderObserver.onChanged() -> requestRecentFolders()

In this critical region, recent folders have been enabled, but we
might not have a recent folder list.

The best solution is to request recent folders when we
enableRecentFolders() to ensure the freshest list.

Bug: 7311426 Recent labels missing

Change-Id: I6116fe5ba68d9c8f6a0ac2c3f803a9f9502140aa