Finesse recent folders requests.
Improves upon the previous CL, by only requesting recent folders if
they are missing. This should improve performance on tablet, where we
frequently enable/disable recent folders as we switch to conversation
mode and back to conversation list mode.

It continues to fix the bug the previous CL sought to fix.

Previous CL:
https://googleplex-android-review.googlesource.com/#/c/237740/

Bug: 7311426 Recent labels missing

Change-Id: I5f9c09e2ef186c7a96bec2e4757d025d2c78c607