use hardware layer to improve conversation load animation fps
Follow the standard pattern to enable HW acceleration when
animating alpha.
Improves framerate of the fade-in on older hardware.

Also add buildLayer() to the pattern (copied from Launcher and
QuickContacts) to do some work up-front instead of during the
animation.

Fix the multi-dismiss case in swipe to switch back to no layer
when the animation is complete.

Bug: 7360692
Change-Id: Id97c473eeffdabe15e3308aa4f1e0755c3f33019