Activated text span should be static.
The SpannableStringBuilder accepts any Object as a span; the only way
to compare objects is through pointer comparison.  If the activated
text span is an instance variable, we create a new instance of it for
each ConvItemView when the activity is recreated through the normal
activity life-cycle.

Making the activated text span static avoids these problems, and
guarantees that we only create a single one per process instance.

Bug: 7316164 Shows half white color sender name in Gmail conversation

Change-Id: I4c650ecd6a396547bbd13ec4a9253433ebd493a8