Decrement the deleted count.
If the item was deleted, and its gone from the cursor being synced,too,
decrement the count regardless.

Only remove it from the iterator if it was explicityly deleted by the user
or within the time window.

Fixes b/7409999 Gmail seems to be losing the last few rows of messages when archiving
tested against mark's case

Change-Id: I943fc9a40a534b53b0a7bfbbf0986241a07f7967