add sanity checks when updating snap header in-place
If snap index points to a collapsed header, mSnapHeader will be
unbound. It's invalid to ask the snap header to refresh in this
state, so only refresh if it is bound to an affected item.

In practice, this really doesn't happen much, because you'd
have to send a message, collapse it, then scroll down far enough
(perhaps you send another message) so the sending collapsed
header intersects the scroll top.

Prevent header view rendering on null for good measure.

Bug: 7459484
Change-Id: I3bf4af97102a0f6b1af5405ffdac72f26e5c7620