speculative fix for native WebView crashes
Towards the end of MR1, we saw more native crashes on
conversation load. It's possible it was purely due to increased
testing, but just in case it was due to loading animation
changes, go back to Gmail1-style content-ready code.

The document has a small div that initially has an
"initial-load" class set on JB+. The class is no longer
dynamically added in JS. The animation-start event listener
remains as-is, configured on DOM parse (and not later upon DOM
load).

Give the animation a different name from the CSS class that
triggers it, on the off-chance that this is crossing some wires.

Bug: 7370125
Change-Id: I4f0f807fbb9e32e024b6c74a6d9684e7fb337695