fix starring after rotation
Starring a message depended on having a working
ConversationController. But while rotation retains the
Loader/Cursor/Messages, it makes the controller (actually a
Fragment) useless because a new one is made.

So have onLoadFinished ALWAYS update the Cursor's controller
to point to the current Fragment before using it.

Similarly, it isn't great that the Cursor points to a
Conversation, which may be old after rotation. That one is less
bad, since a Conversation is relatively static, but everybody
should be on the same page after rotation, so stop caching
internal Conversation references in the Message/Cursor/Loader.

No null checks on conversation/controller get(), because they
really should never be null on access. NPE is as good as
anything to catch these cases.

Bug: 7295695
Change-Id: I949f53779afc820eae0f05da4618e570db381942