Remove leave behind after committing
If we do not null out the leave behind, successive calls to
commitLeaveBehind will try to commit it. For cases where this is an
archive operation, the leave behind item will not have mMostlyDead set
to true, and it will cause the conversation to be deleted entirely.
This is bad.

[It is a separate bug that commitLeaveBehind is called repeatedly,
sometimes 10+ times, while building the MailSpinner.  That's a
separate fix.]

When we commit a leave behind, we should null it out.

Bug: 7525692 ListPopupWindow is committing the LeaveBehindItem in an
             inconsistent way

Change-Id: Ie04563a257da44e770a41006388b080de99f150a