Call commitLeaveBehind only once from the spinner
Building a ListPopupWindow calls getView(position = 0) repeatedly.
This has the effect of calling commitLeaveBehind many times.  This
exposed the problems with commitLeaveBehind that are being fixed
separately.  While it should be safe to call commitLeaveBehind
repeatedly, there is no reason to.  So calling this just once for the
account spinner.  Should also help with Chet's bug on slow
instantiation of the spinner.

Bug: 7416297 long delay after clicking on account/label switch box

Bug: 7525692 ListPopupWindow is committing the LeaveBehindItem in an
             inconsistent way

Change-Id: I872bbe876d5393b194666c39ec3a1662f9f995dc