Make variables final, reduce scope overall
Get the authority and the operation list from the batchMap to allow the
batchMap to go out of scope at the end of the method.

Make the authority final in the for to allow lexical closure without
an additional variable.

Save the isUiThread in a boolean so we don't query it repeatedly
inside the loop.

Reduce scope of try to cover ContentResolver.applyBatch()

Change-Id: I1adbce5f2e36d99b26dc240fc0835d62ff89d1f9