Revert to old conversation position calculation
Use the old way to determine the position of a conversatin
(pre https://googleplex-android-review.googlesource.com/#/c/250616)

We can't use the underlying cursor to determine position because
it doesn't know about the deleted items.  Ideally the ConversationCursor
would contain the map of conversations to positions.  Bug b/7543652 has
been written up to track adding this to the conversation cursor

Bug: 7541980
Change-Id: I925956d135d7b8caf73db7d3013b5baa77664f26