fix account ordering upon cold process start
A cold start would use the cache contents without
AccountManager assistance, and the cache was not well-ordered
when persisted. Add positions to the serialized entries and sort
when loading from the cache to fix that.

Another problem was that subsequent partial updates to the list
of accounts caused the updated entries to fall to the bottom of
the ordering. So change it so incomplete cursor updates use any
existing entry position. Complete cursor updates will still add
to the end, effectively re-doing the ordering as before.

Bug: 7100404
Change-Id: I6e07826c51d013d5954fa39e13cc20410c595c1d