more column annotation
Generate the ordered ACCOUNTS_PROJECTION array from the key set
of a richer map that annotates each account column with its
expected type (typically one of Integer or String).

This is a tiny step towards more unit tests. In this case, I
want to synthesize an Account cursor and fill columns that I
don't care about with dummy values.

Change-Id: If1c82af8575b11fb160a509a99b3ff264408a719